@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:clock/clock.dart';
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:diac_client/src/diac_client.dart';
 import 'package:diac_client/src/diac_event.dart';
 import 'package:diac_client/src/dto/diac_dto.dart';
@@ -8,7 +9,6 @@ import 'package:diac_client/src/dto/diac_store.dart';
 import 'package:expressions/expressions.dart';
 import 'package:flutter_async_utils/flutter_async_utils.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -16,10 +16,9 @@ final _logger = Logger('diac.diac_bloc');
 
 class _NewData {
   _NewData({
-    @required this.closedMessages,
-    @required this.data,
-  })  : assert(closedMessages != null),
-        assert(data != null);
+    required this.closedMessages,
+    required this.data,
+  });
   final Set<String> closedMessages;
   final DiacData data;
 }
@@ -30,11 +29,10 @@ typedef CustomActionHandler = Future<bool> Function(
 
 class DiacBloc with StreamSubscriberBase {
   DiacBloc({
-    @required DiacOpts opts,
+    required DiacOpts opts,
     this.contextBuilder,
     this.customActions = const {},
-  })  : assert(opts != null),
-        _client = DiacClient(opts: opts) {
+  }) : _client = DiacClient(opts: opts) {
 //    handle(_client.store.onValueChanged.listen((event) {
 //      seenMessages.clear();
 //      seenMessages.addAll(event.seen.map((e) => e.messageUuid));
@@ -58,13 +56,13 @@ class DiacBloc with StreamSubscriberBase {
       _logger.fine('handling event $event');
       if (event is DiacEventDismissed) {
         _logger.finer('setting ${event.message.uuid} as dismissed.');
-        _client.store.update((data) => data.copyWith(seen: [
+        _client.store.update((data) => data!.copyWith(seen: [
               ...data.seen,
               DiacHistory(
                 messageUuid: event.message.uuid,
                 messageKey: event.message.key,
                 closedAt: clock.now(),
-                action: event.action.key,
+                action: event.action!.key,
               ),
             ]));
       }
@@ -74,13 +72,13 @@ class DiacBloc with StreamSubscriberBase {
   final DiacClient _client;
   final StreamController<DiacEvent> _events = StreamController.broadcast();
   final Map<String, CustomActionHandler> customActions;
-  Stream<_NewData> _seenMessages;
+  late Stream<_NewData> _seenMessages;
 
   Stream<DiacEvent> get events => _events.stream;
 
   /// callback to create additional context data when evaluating the
   /// expression of [DiacMessage].
-  final AdditionalContextBuilder contextBuilder;
+  final AdditionalContextBuilder? contextBuilder;
 
   final Map<String, ValueStream<DiacMessageDisplay>> _messageForLabel = {};
 
@@ -98,19 +96,21 @@ class DiacBloc with StreamSubscriberBase {
               },
             ),
           )
+          .where((event) => event != null)
+          .cast<DiacMessageDisplay>()
           .publishValueAsync()
           .autoConnect();
 
-  Future<DiacMessageDisplay> _findNextMessageFromData(
+  Future<DiacMessageDisplay?> _findNextMessageFromData(
     DiacData data,
     Set<String> seenMessages,
     Map<String, Object> context,
   ) async {
     try {
       final now = clock.now();
-      _logger.finest('Find next message. ${data.lastConfig.messages}');
+      _logger.finest('Find next message. ${data.lastConfig!.messages}');
       final exprContext = await _createExpressionContext(data, context);
-      for (final message in data.lastConfig.messages) {
+      for (final message in data.lastConfig!.messages) {
         if (seenMessages.contains(message.uuid)) {
           _logger.finest('message was already seen ${message.uuid}');
           continue;
@@ -119,7 +119,7 @@ class DiacBloc with StreamSubscriberBase {
           continue;
         }
         if (message.expression != null) {
-          final expr = Expression.parse(message.expression);
+          final expr = Expression.parse(message.expression!);
           const evaluator = MapAwareEvaluator();
           try {
             final dynamic result = evaluator.eval(
@@ -155,8 +155,6 @@ class DiacBloc with StreamSubscriberBase {
 
   Future<Map<String, Object>> _createExpressionContext(
       DiacData data, Map<String, Object> context) async {
-    assert(data != null);
-    assert(context != null);
     final days = data.firstLaunch.difference(clock.now()).abs().inDays;
     final pi = await _client.opts.getPackageInfo();
     return {
@@ -170,27 +168,25 @@ class DiacBloc with StreamSubscriberBase {
       },
       'action': (String messageKey) {
         final result = data.seen
-            .lastWhere(
-                (element) =>
-                    element.messageKey == messageKey && element.action != null,
-                orElse: () => null)
+            .lastWhereOrNull((element) =>
+                element.messageKey == messageKey && element.action != null)
             ?.action;
         _logger.fine('finding action for $messageKey = $result');
         return result;
       },
-      ...?contextBuilder == null ? null : await contextBuilder(),
+      ...?contextBuilder == null ? null : await contextBuilder!(),
       ...context,
     };
   }
 
   Future<void> triggerMessageAction({
-    @required DiacMessageDisplay message,
-    @required DiacMessageAction action,
+    required DiacMessageDisplay message,
+    required DiacMessageAction action,
   }) async {
-    Uri uri;
+    Uri? uri;
     if (action.expression != null) {
       final result = await _evaluateMessageAction(
-          expression: action.expression,
+          expression: action.expression!,
           expressionContext: message.expressionContext);
       _logger.finer('action expression result: ${result.runtimeType}: $result');
       if (result is Uri) {
@@ -203,7 +199,7 @@ class DiacBloc with StreamSubscriberBase {
       }
     }
     if (uri == null && action.url != null) {
-      uri = Uri.parse(action.url);
+      uri = Uri.parse(action.url!);
     }
     if (uri != null) {
       await triggerAction(DiacEventTriggerCustom(
@@ -222,8 +218,9 @@ class DiacBloc with StreamSubscriberBase {
     return value?.toString();
   }
 
-  Future<Object> _evaluateMessageAction(
-      {String expression, Map<String, dynamic> expressionContext}) async {
+  Future<Object?> _evaluateMessageAction(
+      {required String expression,
+      required Map<String, dynamic> expressionContext}) async {
     try {
       final expr = Expression.parse(expression.trim());
       const evaluator = MapAwareEvaluator();
@@ -232,8 +229,8 @@ class DiacBloc with StreamSubscriberBase {
         'url': (String baseUri, Map<dynamic, dynamic> queryParameters) {
           final uri = Uri.parse(baseUri);
           return uri.replace(queryParameters: <String, dynamic>{
-            ...?uri.queryParametersAll,
-            ...?queryParameters.map<String, dynamic>(
+            ...uri.queryParametersAll,
+            ...queryParameters.map<String, dynamic>(
                 (dynamic key, dynamic value) => MapEntry<String, dynamic>(
                     key.toString(), _asStringOrIterable(value))),
           });
@@ -297,7 +294,7 @@ extension DiacBlocExt on DiacBloc {
 }
 
 extension on DateTime {
-  bool isInRange(DateTime start, DateTime end) =>
+  bool isInRange(DateTime? start, DateTime? end) =>
       (start == null || isAfter(start)) && (end == null || isBefore(end));
 }
 
@@ -325,9 +322,9 @@ class MapAwareEvaluator extends ExpressionEvaluator {
 
 class DiacMessageDisplay {
   DiacMessageDisplay({
-    @required this.message,
-    @required this.expressionContext,
-  }) : assert(message != null);
+    required this.message,
+    required this.expressionContext,
+  });
   final DiacMessage message;
   final Map<String, dynamic> expressionContext;
 }
